split readarr into more files
